Add files via upload
This paper is focused on the development of a model for achieving optimal control of the cooling system of a polymer electrolyte membrane fuel cell (PEMFC)-
based cogeneration system. This model is developed to help facilitate the development and application of control strategies to maximize the energy efficiencies of
PEMFCs, so that the costs associated with electric and thermal generation can be reduced. The results of experimental analysis conducted using an actual PEMFCbased
combined heat and power system that can produce 600 W of electrical power are presented. Then, the development and validation of a simulation model
of the experimental system are discussed. This model is based on a combination of an artificial neural network (ANN) with a non-linear autoregressive exogenous
configuration and a 3D lookup table (LUT) that updates the data input into the ANN as a function of the electrical power demand and the flow rate and input
temperature of the coolant fluid. Due to the nonlinearity of the data contained in the 3D LUT, an algorithm based on linear interpolation and shape-preserving
piecewise cubic Hermite dynamic functions is implemented to interpolate the data in 3D. As a result, the model can predict the outlet temperature of the coolant
fluid and hydrogen consumption rate of the PEMFC as functions of the inlet temperature and flow rate of the coolant fluid and the electrical power demand. The
proposed model exhibits high accuracy and can be used as a black box for the development of new optimization strategies.